Adjusted Routes.tsx to work with Outlet component
